adding nested NavigationGraph
