+++ app/src/main/java/com/example/spedotransferapp/ui/theme/screens/Sign_In.kt
@@ -180,7 +180,8 @@ fun SignIn(
 
         Button(
             onClick = {
-                viewModel.signInUser(text, password)
+//                viewModel.signInUser(text, password)
+                navController.navigate(Routes.SCAFFOLDSCREEN)
             },
             enabled = true,
             modifier = Modifier
